+++ components/organisms/Header/Header.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import { Link } from 'routes';
+import { PAGE_WITH_GRAPHQL } from 'config/routes';
+
 import Logo from 'components/atoms/Logo';
 
 const HeaderWrapper = styled.header`
@@ -22,7 +25,9 @@ const Header = () => {
   return (
     <HeaderWrapper>
       <Logo />
-      <Profile src={`${process.env.ASSET_HOST}/images/avatar-placeholder.png`} alt="avatar" />
+      <Link route={PAGE_WITH_GRAPHQL.pattern}>
+        <Profile src={`${process.env.ASSET_HOST}/images/avatar-placeholder.png`} alt="avatar" />
+      </Link>
     </HeaderWrapper>
   );
 };
