+++ lib/auth/withAuthSecurity.js
@@ -23,18 +23,17 @@ export default Page =>
             mutation: updateToken,
           });
 
-          console.warn(data);
-
           const {
             data: {
               updateToken: { refreshToken },
             },
+            errors,
           } = data;
 
-          if (refreshToken) {
-            if (isServer) {
-              setRefreshToken({ refreshToken, res });
-            }
+          if (!refreshToken) throw new Error(errors[0].message);
+
+          if (isServer) {
+            setRefreshToken({ refreshToken, res });
           }
         } catch (error) {
           console.error('error', error);
