+++ lib/withApolloClient.js
@@ -8,7 +8,7 @@ import { ApolloProvider } from '@apollo/react-hooks';
 import fetch from 'isomorphic-unfetch';
 
 import { createConsoleLink, createAuthHeaderLink, createUpdateTokenLink } from 'lib/apollo/links';
-import { getAccessToken, setAccessToken, parseJWT } from 'lib/auth/tokens';
+import { getAccessToken, setAccessToken } from 'lib/auth/tokens';
 
 import { PORT, GRAPHQL_APP_URL } from 'config/vars';
 
@@ -85,17 +85,6 @@ function initApolloClient(...args) {
 
 const initOnContext = ctx => {
   const inAppContext = Boolean(ctx.ctx);
-
-  // We consider installing `withApollo({ ssr: true })` on global App level
-  // as antipattern since it disables project wide Automatic Static Optimization.
-  if (process.env.NODE_ENV === 'development') {
-    if (inAppContext) {
-      console.warn(
-        'Warning: You have opted-out of Automatic Static Optimization due to `withApollo` in `pages/_app`.\n' +
-          'Read more: https://err.sh/next.js/opt-out-auto-static-optimization\n',
-      );
-    }
-  }
 
   // Initialize ApolloClient if not already done
   const apolloClient = ctx.apolloClient || initApolloClient(ctx.apolloState || {}, inAppContext ? ctx.ctx : ctx);
